// front/yourCarYourWay/src/app/core/interfaces/user.interface.ts
export interface User {
    id: number;
    gender: string;
    lastName: string;
    firstName: string;
    email: string;
    address: string;
    phone: string;
    password: string;
    createdAt: string;
    lastLoginAt: string;
    accountType: string;
    role: string;
    availability: boolean;
}